Separate out frontend.zmqterminal package.
