+++ IPython/frontend/zmqterminal/kernelmanager.py
